fix tsconfig znd component package
