+++ apps/mobile/babel.config.js
@@ -1,15 +1,33 @@
 module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo',
-      // "module:metro-react-native-babel-preset"
-    ],
-    plugins: ['@babel/plugin-proposal-export-namespace-from', 'react-native-reanimated/plugin',
-      'module-resolver',
-      // ["@babel/plugin-transform-runtime", {
-      //   "helpers": true,
-      //   "regenerator": true
-      // }]
+    presets: ['babel-preset-expo'],
+    plugins: [
+      [
+        'module-resolver',
+        {
+          root: ['.'],
+          alias: {
+            'pixel_component': '../../packages/pixel_component',
+            'afk_nostr_sdk': '../../packages/afk_nostr_sdk/src',
+            'common': '../../packages/common/src',
+            'pixel_ui': '../../packages/pixel_ui/src',
+          },
+          extensions: [
+            '.ios.ts',
+            '.android.ts',
+            '.ts',
+            '.ios.tsx',
+            '.android.tsx',
+            '.tsx',
+            '.jsx',
+            '.js',
+            '.json',
+          ],
+        },
+      ],
+      'react-native-reanimated/plugin',
+      "@babel/plugin-proposal-export-namespace-from",
     ],
   };
 };
