+++ apps/mobile/src/modules/CashuWallet/components/Send/index.tsx
@@ -2,7 +2,7 @@
 import '../../../../../applyGlobalPolyfills';
 
 import {Picker} from '@react-native-picker/picker';
-import {MintData} from 'afk_nostr_sdk/src/hooks/cashu/useCashu';
+import {MintData} from 'afk_nostr_sdk';
 import * as Clipboard from 'expo-clipboard';
 import {randomUUID} from 'expo-crypto';
 import React, {useEffect, useState} from 'react';
