+++ apps/mobile/src/screens/LaunchToken/index.tsx
@@ -2,7 +2,7 @@ import {useNavigation} from '@react-navigation/native';
 import {useQueryClient} from '@tanstack/react-query';
 // import { useAuth } from '../../store/auth';
 import {useAuth, useCreateChannel, useProfile, useSettingsStore} from 'afk_nostr_sdk';
-import {AFK_RELAYS} from 'afk_nostr_sdk/src/utils/relay';
+import {AFK_RELAYS} from 'afk_nostr_sdk';
 import * as Clipboard from 'expo-clipboard';
 import * as ImagePicker from 'expo-image-picker';
 import {Formik, FormikProps} from 'formik';
