+++ packages/afk_nostr_sdk/src/index.ts
@@ -3,3 +3,4 @@ export * from './hooks';
 export * from './storage';
 export * from './store';
 export * from './types';
+export * from './utils';
