--- packages/pixel_component/index.ts
@@ -1,2 +1,0 @@
-export * from "./src/components"
-export * from "./src/layout"
