+++ packages/pixel_component/src/components/canvas/controller.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef, createRef } from "react";
+import React, { useState, useEffect, useRef, createRef } from "react";
 
 import { StencilCreationOverlay } from "./stencil-overlay";
 import { playShutter } from "../utils/sounds";
