+++ packages/pixel_component/src/components/tabs/basic.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { playSoftClick2 } from "../utils/sounds";
 
 export const BasicTab = (props: any) => {
