+++ packages/pixel_component/src/components/tabs/index.tsx
@@ -1,1 +1,4 @@
 export { PixelInfoTab, } from "./pixel"   
+export { ExpandableTab } from "./expandable"
+export { BasicTab } from "./basic"
+export { AccountTab } from "./account"
