+++ packages/pixel_component/src/components/tabs/leaderboard.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import { lookupAddresses } from '@cartridge/controller';
 import { useAccount } from '@starknet-react/core';
 import { BasicTab } from "./basic";
